pyproject.toml
[project]
name = "apiflask-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apiflask==3.0.2",
    "bcrypt==5.0.0",
    "debugpy==1.8.17",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "shortuuid==1.0.13",
    "uwsgi==2.0.31",
]

[dependency-groups]
dev = [
    "debugpy==1.8.17",
]
